Split jdk internal loader substitutions
